daily_planner/pyproject.toml
[project]
name = "daily-planner"
version = "0.1.0"
description = "Planner to help you start your day"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "llama-cpp-agent>=0.2.35",
    "llama-cpp-python>=0.3.6",
    "pre-commit>=4.1.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.3",
]
